+++ test/directives.js
@@ -10,7 +10,7 @@ var html_fixture = fs.readFileSync(__dirname + '/doc.html', {encoding: 'utf8'}),
 
 describe('directives', function() {
 
-    describe('def directives', function() {
+    describe('def directive', function() {
 
 
         it('executes a sub-context', function() {
@@ -127,6 +127,117 @@ describe('directives', function() {
             data.parent_scope.value.should.eql('first nav item');
             data.shadowed.value.should.eql('second nav item');
         });
+    });
 
+    describe('namespace directives', function() {
+
+
+        it('save into a child object', function() {
+            var tt = new TakeTemplate([
+                    'namespace               : parent',
+                    '    $ h1 | 0 text ;         : value'
+                ]),
+                data = tt.take(html_fixture);
+            data.parent.value.should.eql('Text in h1');
+        });
+
+
+        it('has "+" as an alias', function() {
+            var tt = new TakeTemplate([
+                    '+                       : parent',
+                    '    $ h1 | 0 text ;         : value'
+                ]),
+                data = tt.take(html_fixture);
+            data.parent.value.should.eql('Text in h1');
+        });
+
+
+        it('can be deeply nested (namespace inception)', function() {
+            var tt = new TakeTemplate([
+                    '+                       : p0',
+                    '    +                       : p1',
+                    '        +                       : p2',
+                    '            +                       : p3',
+                    '                +                       : p4',
+                    '                    $ h1 | 0 text ;         : value'
+                ]),
+                data = tt.take(html_fixture);
+            data.p0.p1.p2.p3.p4.value.should.eql('Text in h1');
+        });
+
+
+        it('nest and exit correctly', function() {
+            var tt = new TakeTemplate([
+                    '+                               : p0a',
+                    '    namespace                       : p1a',
+                    '        $ a | 0 text ;                  : first_li',
+                    '    +                               : p1b',
+                    '        $ a | 1 text ;                  : second_li',
+                    'namespace                       : p0b',
+                    '    $ a | 2 text ;                  : third_li'
+                ]),
+                data = tt.take(html_fixture);
+            data.p0a.p1a.first_li.should.eql('first nav item');
+            data.p0a.p1b.second_li.should.eql('second nav item');
+            data.p0b.third_li.should.eql('first content link');
+        });
+
+
+        it('can be defined inline', function() {
+            var tt = new TakeTemplate([
+                    '$ h1 ; +                            : parent',
+                    '    | 0 text ;                          : value',
+                ]),
+                data = tt.take(html_fixture);
+            data.parent.value.should.eql('Text in h1');
+        });
+
+
+        it('nest and exit correctly when applied inline', function() {
+            var tt = new TakeTemplate([
+                    '$ a ; +                         : p0a',
+                    '    | 0 ; +                         : p1a',
+                    '        | text ;                        : first_li',
+                    '    | 1 text ;                      : second_li',
+                    '+                               : p0b',
+                    '    $ h1 | 0 text ; +               : p1b',
+                    '                                        : h1'
+                ]),
+                data = tt.take(html_fixture);
+            data.p0a.p1a.first_li.should.eql('first nav item');
+            data.p0a.second_li.should.eql('second nav item');
+            data.p0b.p1b.h1.should.eql('Text in h1');
+        });
+
+
+        it('require minimal indentation', function() {
+            var tt = new TakeTemplate([
+                    '$ a ; +                         : p0a',
+                    ' | 0 ; +                            : p1a',
+                    '  | text ;                              : first_li',
+                    ' | 1 text ;                         : second_li',
+                    ' | 1 text ;                         : second_li_again',
+                    '+                               : p0b',
+                    ' $ h1 | 0 text ; +                  : p1b',
+                    '                                        : h1'
+                ]),
+                data = tt.take(html_fixture);
+            data.p0a.p1a.first_li.should.eql('first nav item');
+            data.p0a.second_li.should.eql('second nav item');
+            data.p0b.p1b.h1.should.eql('Text in h1');
+        });
+
+
+        it('do not overwrite if repeated', function() {
+            var tt = new TakeTemplate([
+                    '+                       : links',
+                    '    $ a | 0 text ;          : first',
+                    '+                       : links',
+                    '    $ a | 1 text ;          : second',
+                ]),
+                data = tt.take(html_fixture);
+            data.links.first.should.eql('first nav item');
+            data.links.second.should.eql('second nav item');
+        });
     });
 });